Move new file to components/wca
